requirements (1).txt
tensorflow
opencv-python
tqdm
numpy
scikit-learn
matplotlib
seaborn
imbalanced-learn
Pillow
reportlab
fpdf
streamlit
